atomic linked list

+++ kcore/atomic/src/host_funcs.cu
@@ -29,7 +29,7 @@ void find_kcore(string data_file,bool write_to_disk){
     unsigned int level = 0;
     unsigned int count = 0;
     unsigned int* global_count  = NULL;
-    unsigned int* blockCounter  = NULL;
+    unsigned long long int* blockCounter  = NULL;
     unsigned int* glBuffers     = NULL;
 
     chkerr(cudaMalloc(&global_count, sizeof(unsigned int)));
